[#62] refactor: change common component import path
